+++ leather/scales/__init__.py
@@ -3,5 +3,6 @@
 from leather.scales.annual import Annual
 from leather.scales.base import Scale
 from leather.scales.linear import Linear
+from leather.scales.monthly import Monthly
 from leather.scales.ordinal import Ordinal
 from leather.scales.temporal import Temporal